+++ CHANGELOG.md
@@ -13,6 +13,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Removed
 
+## [6.1.4] - 2026-05-27
+
+### Changed
+- Added PreCreatingAsync virtual method for running tasks before the container creates
+- PreBuild switched to PreBuildingAsync
+
 ## [6.1.3] - 2026-05-27
 
 ### Fixed